bin/update: Keep previous msgids
Now old msgids are not dropped and so also older releases
of the respective product are still translated. Might get
messy in the future though, since there is no clean up.

But there are not much changes anyway to existing messages,
so let's see how worse this can really get…